fix: Add JIT warm-up to Fibonacci to fix 89% variance
Root cause: JIT compiler optimizes fibonacci loop after first run,
causing 2x performance difference between consecutive runs.

Fix: Add 10,000 iteration warm-up before timed benchmark to force
JIT compilation BEFORE timing starts.

Changes:
- SingleCoreBenchmarks.fibonacciRecursive(): Added JIT warm-up
- MultiCoreBenchmarks.fibonacciRecursive(): Added JIT warm-up
- BenchmarkEvent.kt: monteCarloSamples Int -> Long
- KotlinBenchmarkManager.kt: Added L suffix to Long literals

Expected results:
- SC Fibonacci variance: 89% -> <10%
- MC Fibonacci variance: 30% -> <10%

+++ app/src/main/java/com/ivarna/finalbenchmark2/cpuBenchmark/BenchmarkEvent.kt
@@ -57,7 +57,7 @@ data class WorkloadParams(
         val compressionDataSizeMb: Int = 2,
         val compressionIterations: Int =
                 100, // FIXED WORK PER CORE: Target ~1.5-2.0 seconds execution
-        val monteCarloSamples: Int = 1000000,
+        val monteCarloSamples: Long = 1000000L,
         val jsonDataSizeMb: Int = 1,
         val jsonParsingIterations: Int = 1000, // CACHE-RESIDENT: Number of parsing iterations
         val nqueensSize: Int = 10
